Вынесение логики работы загрузки/сохранения контактов
